Rename: info 파일명 변경
