+++ src/features/info/info.css
